requirements.txt
psycopg2==2.8.4
Scrapy==1.8.0
SQLAlchemy==1.3.13
python-dateutil
nltk
kafka-python